instahyre/requirements.txt
Django==4.2.14
djangorestframework==3.15.2